Create CTEs
Understanding how to use CTEs and how it simplify a query. CTE works as a virtual table with records and columns that is created during the execution of the query. You would use CTEs to create a temporary relation by using the With Clause. Also that is beneficial is that CTEs are more readable than Subqueries.